+++ src/main/java/net/frozenblock/wilderwild/config/WWWorldgenConfig.java
@@ -53,6 +53,7 @@ public final class WWWorldgenConfig {
 				BIRCH_BRANCHES = config.treeGeneration.birchBranches;
 				OAK_BRANCHES = config.treeGeneration.oakBranches;
 				DARK_OAK_BRANCHES = config.treeGeneration.darkOakBranches;
+				PALE_OAK_BRANCHES = config.treeGeneration.paleOakBranches;
 				NEW_WITCH_HUTS = config.structure.newWitchHuts;
 			}
 		}
@@ -62,6 +63,7 @@ public final class WWWorldgenConfig {
 	public static volatile boolean BIRCH_BRANCHES = true;
 	public static volatile boolean OAK_BRANCHES = true;
 	public static volatile boolean DARK_OAK_BRANCHES = true;
+	public static volatile boolean PALE_OAK_BRANCHES = true;
 	public static volatile boolean NEW_WITCH_HUTS = true;
 
 	@CollapsibleObject
@@ -133,6 +135,9 @@ public final class WWWorldgenConfig {
 		@EntrySyncData("mushroomGeneration")
 		public boolean mushroomGeneration = true;
 
+		@EntrySyncData("paleMushroomGeneration")
+		public boolean paleMushroomGeneration = true;
+
 		@EntrySyncData("pollen")
 		public boolean pollen = true;
 
@@ -167,6 +172,12 @@ public final class WWWorldgenConfig {
 
 		@EntrySyncData("mossDecoration")
 		public boolean mossDecoration = true;
+
+		@EntrySyncData("redMossDecoration")
+		public boolean redMossDecoration = true;
+
+		@EntrySyncData("paleMossDecoration")
+		public boolean paleMossDecoration = true;
 
 		@EntrySyncData("scorchedSandDecoration")
 		public boolean scorchedSandDecoration = true;
@@ -263,6 +274,9 @@ public final class WWWorldgenConfig {
 
 		@EntrySyncData("darkOakBranches")
 		public boolean darkOakBranches = true;
+
+		@EntrySyncData("paleOakBranches")
+		public boolean paleOakBranches = true;
 	}
 
 	public static class AquaticGeneration {
@@ -301,6 +315,9 @@ public final class WWWorldgenConfig {
 
 		@EntrySyncData("oceanMossGeneration")
 		public boolean oceanMossGeneration = true;
+
+		@EntrySyncData("oceanRedMossGeneration")
+		public boolean oceanRedMossGeneration = true;
 	}
 
 	public static class BiomePlacement {